refactor(bottomMenuDialog): 하단 메뉴 다이얼로그 관련 클래스들을 패키지화 및 MenuItemType 구현
