+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/common/views/bottomMenuDialog/MenuItemType.kt
@@ -1,0 +1,5 @@
+package com.emmsale.presentation.common.views.bottomMenuDialog
+
+enum class MenuItemType {
+    NORMAL, DANGER
+}
