+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/eventdetail/comment/recyclerView/CommentViewHolder.kt
@@ -5,9 +5,9 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.emmsale.R
 import com.emmsale.databinding.ItemCommentsCommentsBinding
-import com.emmsale.presentation.common.views.BottomDialogMenuItem
-import com.emmsale.presentation.common.views.BottomMenuDialog
 import com.emmsale.presentation.common.views.DangerDialog
+import com.emmsale.presentation.common.views.bottomMenuDialog.BottomMenuDialog
+import com.emmsale.presentation.common.views.bottomMenuDialog.MenuItemType
 import com.emmsale.presentation.ui.eventdetail.comment.uiState.CommentUiState
 
 class CommentViewHolder(
@@ -62,7 +62,7 @@ class CommentViewHolder(
     private fun BottomMenuDialog.addReportButton() {
         addMenuItemBelow(
             context.getString(R.string.all_reportButtonLabel),
-            BottomDialogMenuItem.DANGER,
+            MenuItemType.DANGER,
         ) { }
     }
